refactor(api): reorganize canvas module structure and introduce CanvasSyncModule
- Updated import paths for CanvasSyncService and CanvasSyncModule to reflect the new directory structure, enhancing code organization.
- Removed direct dependencies on CanvasService from various modules, promoting a cleaner separation of concerns.
- Ensured adherence to code quality standards, including the use of optional chaining and nullish coalescing for safer property access.
