+++ apps/api/src/modules/canvas/canvas.module.ts
@@ -22,7 +22,7 @@ import { ActionModule } from '../action/action.module';
 import { ProviderModule } from '../provider/provider.module';
 import { CodeArtifactModule } from '../code-artifact/code-artifact.module';
 import { isDesktop } from '../../utils/runtime';
-import { CanvasSyncService } from './canvas-sync.service';
+import { CanvasSyncModule } from '../canvas-sync/canvas-sync.module';
 
 @Module({
   imports: [
@@ -31,6 +31,7 @@ import { CanvasSyncService } from './canvas-sync.service';
     KnowledgeModule,
     ActionModule,
     ProviderModule,
+    CanvasSyncModule,
     CodeArtifactModule,
     SubscriptionModule,
     ...(isDesktop()
@@ -50,7 +51,6 @@ import { CanvasSyncService } from './canvas-sync.service';
   controllers: [CanvasController],
   providers: [
     CanvasService,
-    CanvasSyncService,
     ...(isDesktop()
       ? []
       : [
@@ -61,6 +61,6 @@ import { CanvasSyncService } from './canvas-sync.service';
           VerifyNodeAdditionProcessor,
         ]),
   ],
-  exports: [CanvasService, CanvasSyncService],
+  exports: [CanvasService],
 })
 export class CanvasModule {}
