+++ apps/api/src/modules/canvas/canvas.service.ts
@@ -19,7 +19,6 @@ import {
   User,
   SkillContext,
   ActionResult,
-  CanvasNode,
 } from '@refly/openapi-schema';
 import { Prisma } from '../../generated/client';
 import { genCanvasID, genTransactionId } from '@refly/utils';
@@ -35,8 +34,8 @@ import { RedisService } from '../common/redis.service';
 import { ObjectStorageService, OSS_INTERNAL } from '../common/object-storage';
 import { ProviderService } from '../provider/provider.service';
 import { isDesktop } from '../../utils/runtime';
-import { CanvasSyncService } from './canvas-sync.service';
-import { CanvasNodeFilter, initEmptyCanvasState, prepareAddNode } from '@refly/canvas-common';
+import { CanvasSyncService } from '../canvas-sync/canvas-sync.service';
+import { initEmptyCanvasState } from '@refly/canvas-common';
 
 @Injectable()
 export class CanvasService {
@@ -357,70 +356,6 @@ export class CanvasService {
     });
 
     return canvas;
-  }
-
-  /**
-   * Add a node to the canvas
-   * @param user - The user who is adding the node
-   * @param canvasId - The id of the canvas to add the node to
-   * @param node - The node to add
-   * @param connectTo - The nodes to connect to
-   * @param options - Additional options including autoLayout
-   */
-  /**
-   * Add a node to the canvas
-   * @param user - The user who is adding the node
-   * @param canvasId - The id of the canvas to add the node to
-   * @param node - The node to add (id is optional)
-   * @param connectTo - The nodes to connect to
-   */
-  async addNodeToCanvas(
-    user: User,
-    canvasId: string,
-    node: Pick<CanvasNode, 'type' | 'data'> & Partial<Pick<CanvasNode, 'id'>>,
-    connectTo?: CanvasNodeFilter[],
-    options?: { autoLayout?: boolean },
-  ) {
-    const releaseLock = await this.canvasSyncService.lockState(canvasId);
-    const { nodes, edges } = await this.canvasSyncService.getCanvasData(user, { canvasId });
-
-    this.logger.log(
-      `[addNodeToCanvas] add node to canvas ${canvasId}, node: ${JSON.stringify(node)}, autoLayout: ${options?.autoLayout}`,
-    );
-    const { newNode, newEdges } = prepareAddNode({
-      node,
-      nodes,
-      edges,
-      connectTo,
-      autoLayout: options?.autoLayout, // Pass autoLayout parameter
-    });
-
-    await this.canvasSyncService.syncState(
-      user,
-      {
-        canvasId,
-        transactions: [
-          {
-            txId: genTransactionId(),
-            createdAt: Date.now(),
-            syncedAt: Date.now(),
-            nodeDiffs: [
-              {
-                type: 'add',
-                id: newNode.id,
-                to: newNode,
-              },
-            ],
-            edgeDiffs: newEdges.map((edge) => ({
-              type: 'add',
-              id: edge.id,
-              to: edge,
-            })),
-          },
-        ],
-      },
-      { releaseLock },
-    );
   }
 
   async updateCanvas(user: User, param: UpsertCanvasRequest) {
