+++ apps/api/src/modules/collab/collab.module.ts
@@ -7,13 +7,13 @@ import { QUEUE_SYNC_CANVAS_ENTITY } from '../../utils/const';
 import { CollabService } from './collab.service';
 import { CollabController } from './collab.controller';
 import { isDesktop } from '../../utils/runtime';
-import { CanvasModule } from '../canvas/canvas.module';
+import { CanvasSyncModule } from '../canvas-sync/canvas-sync.module';
 
 @Module({
   imports: [
     CommonModule,
     RAGModule,
-    CanvasModule,
+    CanvasSyncModule,
     ...(isDesktop() ? [] : [BullModule.registerQueue({ name: QUEUE_SYNC_CANVAS_ENTITY })]),
   ],
   providers: [CollabGateway, CollabService],
