+++ apps/api/src/modules/pilot/pilot.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { BullModule } from '@nestjs/bullmq';
 import { CommonModule } from '../common/common.module';
 import { CanvasModule } from '../canvas/canvas.module';
+import { CanvasSyncModule } from '../canvas-sync/canvas-sync.module';
 import { SkillModule } from '../skill/skill.module';
 import { KnowledgeModule } from '../knowledge/knowledge.module';
 import { ProviderModule } from '../provider/provider.module';
@@ -16,6 +17,7 @@ import { QUEUE_RUN_PILOT } from '../../utils/const';
   imports: [
     CommonModule,
     CanvasModule,
+    CanvasSyncModule,
     SkillModule,
     ProviderModule,
     KnowledgeModule,
