+++ apps/api/src/modules/skill/skill-engine.service.ts
@@ -63,7 +63,7 @@ export class SkillEngineService implements OnModuleInit {
         return result;
       },
       generateMedia: async (user, req) => {
-        const result = await this.mediaGeneratorService?.generate?.(user, req);
+        const result = await this.mediaGeneratorService.generate(user, req);
         return result;
       },
       getActionResult: async (user, req) => {
